:recycle: Split the tests
