+++ setup.cfg
@@ -4,19 +4,21 @@ include_package_data = True
 packages = find:
 tests_require =
     tox
-    pytest
+    pytest-cov
     freezegun
     isort
     black
+    flake8
     coverage
 
 [options.extras_require]
 tests =
     tox
-    pytest
+    pytest-cov
     freezegun
     isort
     black
+    flake8
     coverage
 pep8 = flake8
 coverage = coverage
@@ -40,7 +42,7 @@ skip = env,.tox,.history,.eggs
 not_skip = __init__.py
 
 [tool:pytest]
-testpaths = tests.py
+testpaths = tests/
 
 
 [pep8]
