+++ tox.ini
@@ -11,7 +11,7 @@ deps =
     pytest-cov
     freezegun
 commands =
-  py.test tests.py --cov=json_logic --cov-report xml:reports/coverage.xml
+  py.test tests/ --cov=json_logic --cov-report xml:reports/coverage.xml
 
 [testenv:clean]
 deps = coverage
@@ -25,10 +25,10 @@ commands = isort --recursive --check-only --diff .
 [testenv:black]
 deps = black
 skipsdist = True
-commands = black --check json_logic
+commands = black --check .
 
 [testenv:flake8]
 deps = flake8
 basepython = python
-commands = flake8 json_logic
+commands = flake8 .
